requirements.txt
Django==4.0.4
djangorestframework==3.13
django-redis==5.2.0
psycopg2>=2.8
redis==4.2.2